+++ kaye/cli/cli_skill/__init__.py
@@ -1,3 +1,8 @@
+"""
+CLI subcommand for Agent Skill integration.
+"""
+
+
 def convert_display_name2skill_name(display_name):
     return display_name.replace(" ", "-").lower()